# user-service-backup-temp/openapi.yaml
openapi: 3.1.0
info:
  title: Peerit User Service API
  description: User profiles, roles, and team membership management for the Peerit platform
  version: 1.0.0
  contact:
    name: Peerit Team
    url: https://github.com/PXL-Digital-Application-Samples/peerit
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3020
    description: Development server
  - url: https://api.peerit.dev/users
    description: Production server
...
paths:
...
  /api/users/{userId}/teams:
    get:
      tags: [Team Membership]
      summary: Get user's team memberships
      description: Retrieve all teams a user belongs to
      security:
        - keycloakOAuth2: [admin, teacher, student]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team memberships retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMembership'
...
components:
  schemas:
...
    TeamMembership:
      type: object
      properties:
        teamId:
          type: string
          format: uuid
        teamName:
          type: string
        teamRole:
          type: string
          enum: [leader, member]
        joinedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - teamId
        - teamName
        - teamRole
...
  securitySchemes:
    keycloakOAuth2:
      type: oauth2
      description: Keycloak OAuth2 with role-based scopes
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/peerit/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/peerit/protocol/openid-connect/token
          scopes:
            admin: "Full administrative access to all platform features"
            teacher: "Teacher role - can manage courses, teams, and view all reviews"
            student: "Student role - can join teams and submit peer reviews"
        password:
          tokenUrl: http://localhost:8080/realms/peerit/protocol/openid-connect/token
          scopes:
            admin: "Full administrative access to all platform features"
            teacher: "Teacher role - can manage courses, teams, and view all reviews"
            student: "Student role - can join teams and submit peer reviews"
security:
  - keycloakOAuth2: []